% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_classic.R
\docType{data}
\name{apsim.options}
\alias{apsim.options}
\title{Environment which stores APSIM options}
\format{
An object of class \code{environment} of length 3.
}
\usage{
apsim.options
}
\value{
This is an environment, so nothing to return.
}
\description{
Environment which can store the path to the executable and where examples are located.
             Creating an environment avoids the use of global variables or other similar practices
             which would have possible undesriable consequences.
}
\details{
Environment which stores APSIM options
}
\examples{
\dontrun{
names(apsim.options)
apsim_options(exe.path = "some-new-path-to-executable")
apsim.options$exe.path
}

}
\keyword{datasets}
