% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_version.R
\name{apsim_version}
\alias{apsim_version}
\title{Display available APSIM \sQuote{Classic} and APSIM-X versions}
\usage{
apsim_version(
  which = c("all", "inuse"),
  version = c("number", "full"),
  verbose = TRUE
)
}
\arguments{
\item{which}{either \sQuote{all} or \sQuote{inuse}}

\item{version}{either \sQuote{number} for a compact display or \sQuote{full}}

\item{verbose}{whether to print the information to standard output}
}
\value{
a data frame (all) or a vector (inuse) with APSIM-X and/or APSIM versions
}
\description{
Display available APSIM \sQuote{Classic} and APSIM-X versions
}
\examples{
\dontrun{
## Check which apsim version are avialable
ava <- apsim_version(verbose = TRUE)
}

}
