% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\docType{data}
\name{apsimx.options}
\alias{apsimx.options}
\title{Environment which stores APSIM-X options}
\format{An object of class \code{environment} of length 2.}
\usage{
apsimx.options
}
\description{
Environment which can store the path to the executable and where examples are located
}
\details{
Environment which stores APSIM-X options
}
\keyword{datasets}
