% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{apsimx}
\alias{apsimx}
\title{Run an APSIM-X simulation}
\usage{
apsimx(file = "", src.dir = ".", silent = FALSE, value = c("all",
  "report", "none"), cleanup = c(0, 1, 2))
}
\arguments{
\item{file}{file name to be run (the extension .apsimx is optional)}

\item{src.dir}{directory containing the .apsimx file to be run (defaults to the current directory)}

\item{silent}{whether to print messages from apsim executable}

\item{value}{how much output to return:
option 'all' returns all components of the simulation;
option 'report' returns only the 'main' report component;
option 'none' does not create a data.frame but it generates the databases from APSIM-X}

\item{cleanup}{level of file cleanup:
0 does nothing; 
1 deletes the .db created by APSIM-X;
2 is not implemented yet.}
}
\description{
The goal is to run apsimx from R. It basically calls 'system' to run from the command line.
}
\details{
Run an APSIM-X Simulation

A valid apsimx file can be run from within R. The main goal is to make running APSIM-X
simple, especially for large scale simulations or parameter optimization
}
\examples{
\dontrun{
## See function 'apsimx_example' 
}

}
