% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{apsimx_options}
\alias{apsimx_options}
\title{Setting some options for the package}
\usage{
apsimx_options(exe.path = NA, examples.path = NA)
}
\arguments{
\item{exe.path}{path to the executable}

\item{examples.path}{path to the examples}
}
\description{
Used to set the path to the APSIM-X executable.
}
\details{
Set APSIM-X options
}
\note{
In Windows it is more likely that APSIM-X is installed in some alternative location 
       and guessing this can be difficult to auto detect.
}
