% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_classic.R
\name{auto_detect_apsim_examples}
\alias{auto_detect_apsim_examples}
\title{Auto detect where apsim examples are located}
\usage{
auto_detect_apsim_examples()
}
\value{
will create a directory pointing to APSIM \sQuote{Classic} distributed examples
}
\description{
simple function to detect where APSIM \sQuote{Classic} examples are located
}
\details{
Auto detect where APSIM (7.x) \sQuote{Classic} examples are located
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsim_examples()
}

}
