% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{auto_detect_apsimx_examples}
\alias{auto_detect_apsimx_examples}
\title{Auto detect where apsimx examples are located}
\usage{
auto_detect_apsimx_examples()
}
\value{
will create a directory pointing to APSIM-X distributed examples
}
\description{
simple function to detect where APSIM-X examples are located
}
\details{
Auto detect where apsimx examples are located
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsimx_examples()
}

}
