% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{check_apsim_met}
\alias{check_apsim_met}
\title{Check a met file for possible errors}
\usage{
check_apsim_met(met)
}
\arguments{
\item{met}{object of class \sQuote{met}}
}
\value{
does not return anything unless possible errors are found
}
\description{
Takes in an object of class \sQuote{met} and checks for missing/valid/reasonable values
}
\details{
It will only check for missing values and reasonable (within range) values for:
 \sQuote{year}: range (1500 to 3000); \cr
 \sQuote{day}: range (1 to 366); \cr 
 \sQuote{maxt}: range (-60 to 60) -- units (C); \cr
 \sQuote{mint}: range (-60 to 40) -- units (C); \cr
 \sQuote{radn}: range (0 to 40) -- units (MJ/m2/day);  \cr
 \sQuote{rain}: range (0 to 100) -- units (mm/day)
}
