% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conv.R
\name{date2doy}
\alias{date2doy}
\title{lazy version of the inverse of the previous function}
\usage{
date2doy(x, year = 2001, inverse = FALSE)
}
\arguments{
\item{x}{either an integer 1-366 or a \sQuote{Date}}

\item{year}{year}

\item{inverse}{if TRUE it goes from doy to \sQuote{Date}}
}
\description{
Given a \sQuote{Date} it converts to julian day (1-366) or day of the year
}
\examples{
date2doy("04-30")

}
