% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conv.R
\name{doy2date}
\alias{doy2date}
\title{Converts from doy to date}
\usage{
doy2date(x, year = 2001, inverse = FALSE)
}
\arguments{
\item{x}{either an integer 1-366 or a \sQuote{Date}}

\item{year}{year}

\item{inverse}{if TRUE it goes from date to day}
}
\description{
Given a day of the year as julian (1-366) it converts to \sQuote{Date}
}
\details{
This function goes from doy to date
}
\examples{
doy2date(120)
}
