% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsimx.R
\name{edit_apsimx}
\alias{edit_apsimx}
\title{Edit an APSIM-X Simulation}
\usage{
edit_apsimx(file, src.dir = ".", wrt.dir = NULL, node = c("Weather",
  "Soil", "SurfaceOrganicMatter", "MicroClimate", "Crop", "Manager"),
  soil.child = c("Water", "Nitrogen", "OrganicMatter", "Analysis",
  "InitialWater", "Sample"), som.child = c("Pools", "Other"),
  parm = NULL, value = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{file}{file ending in .apsimx to be edited}

\item{src.dir}{directory containing the .apsimx file to be edited; defaults to the current working directory}

\item{wrt.dir}{should be used if the destination directory is different from the src.dir}

\item{node}{either 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop' or 'Manager' (for now)}

\item{soil.child}{specific soil component to be edited}

\item{som.child}{specific soil organic matter component to be edited}

\item{parm}{parameter to be edited}

\item{value}{new values for the parameter to be edited}

\item{overwrite}{logical; if \code{TRUE} the old file is overwritten, a new file is written otherwise}

\item{verbose}{whether to print information about successful edit}
}
\value{
(when verbose=TRUE) complete file path to edited .apsimx file is returned as a character string.
As a side effect this function created a new (XML) .apsimx file.
}
\description{
This function allows editing of an APSIM-X simulation file.
}
\details{
The variables specified by \code{parm} within the .apsimx file specified by \code{file} 
in the source directory \code{src.dir} are edited. The old values are replaced with \code{value}, which
is a list that has the same number of elements as the length of the vector \code{parm}.  The current
.apsimx file will be overwritten if \code{overwrite} is set to \code{TRUE}; otherwise the file
\emph{'file'-edited.apsimx} will be created.  If (verbose = TRUE) then the name
 of the written file is returned. The function is similar to the edit_apsim functin in the 'apsimr'
 package, but with the difference that only some variables (parameters) can be modified.
 
 The function inspect_apsimx is for a quick look from within R. The APSIM GUI provides a more
 complete examination of the .apsimx file
}
\note{
The components that can be edited are restricted becuase this is better in preventing
errors of editing unintended parts of the file. The disadvantage is that there is less flexibility
compared to the similar function in the 'apsimr' package
}
\examples{
\dontrun{
## This example will read one of the examples distributed with APSIM-X
## but write to the current directory

ex.dir <- auto_detect_apsimx_examples()
ocs <- rep(0.5, 7)
edit_apsimx("Maize.apsimx", src.dir = ex.dir,
            wrt.dir = ".",
            node = "Soil",
            soil.child = "OrganicMatter", 
            parm = "OC", value = ocs,
            verbose = FALSE)
## To delete the file run
system("rm Maize-edited.apsimx")
}

}
