% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chirps_apsim_met.R
\name{get_chirps_apsim_met}
\alias{get_chirps_apsim_met}
\title{Get CHIRPS data for an APSIM met file}
\usage{
get_chirps_apsim_met(
  lonlat,
  dates,
  wrt.dir = ".",
  filename = NULL,
  fillin.radn = TRUE,
  silent = FALSE
)
}
\arguments{
\item{lonlat}{Longitude and latitude vector}

\item{dates}{date ranges}

\item{wrt.dir}{write directory}

\item{filename}{file name for writing out to disk}

\item{fillin.radn}{whether to fill in radiation data using the nasapower pacakge. Default is TRUE.}

\item{silent}{defaut is FALSE. Changing it will not do anything at the moment. A future feature.}
}
\value{
returns an object of class \sQuote{met} and writes a file to disk when filename is supplied.
}
\description{
Uses \code{\link[chirps]{get_chirps}} from the \CRANpkg{chirps} package to download data to create an APSIM met file.
}
\details{
This function requires the \CRANpkg{chirps} package

If the filename is not provided it will not write the file to disk,
but it will return an object of class \sQuote{met}. This is useful in case manipulation
is required before writing to disk.
}
\examples{
\dontrun{
require(chirps)
## This will not write a file to disk
chrp <- get_chirps_apsim_met(lonlat = c(-93,42), dates = c("2012-01-01","2012-12-31"))
}

}
