% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power_apsim_met.R
\name{get_power_apsim_met}
\alias{get_power_apsim_met}
\title{Get POWER data for an APSIM met file}
\usage{
get_power_apsim_met(lonlat, dates, wrt.dir = ".", filename = NULL)
}
\arguments{
\item{lonlat}{Longitude and latitude vector}

\item{dates}{date ranges}

\item{wrt.dir}{write directory}

\item{filename}{file name for writing out to disk}
}
\description{
Uses 'get_power' from the 'nasapower' package to download data to create an APSIM met file.
}
\details{
This requires the nasapower package
}
\examples{
\dontrun{
require(nasapower)
## This will not write a file to disk
pwr <- get_power_apsim_met(lonlat = c(-93,42), dates = c("2017-01-01","2017-12-31"))
}



}
