% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worldmodeler_soil_profile.R
\name{get_worldmodeler_soil_profile}
\alias{get_worldmodeler_soil_profile}
\title{Obtain a synthetic APSIM soil profile from the World Modeler database}
\usage{
get_worldmodeler_soil_profile(
  lonlat,
  fix = FALSE,
  check = FALSE,
  soil.name,
  wrt.dir,
  filename,
  verbose = FALSE,
  cleanup = FALSE
)
}
\arguments{
\item{lonlat}{Longitude and latitude vector (e.g. c(-93, 42)) or matrix.}

\item{fix}{whether to fix the soil profiles}

\item{check}{whether to check the soil profiles}

\item{soil.name}{optional soil name}

\item{wrt.dir}{optional directory where to save a file with \sQuote{soils} extension.
If missing it will be written to a temporary directory.}

\item{filename}{optional name to be used when saving the file}

\item{verbose}{verbose argument passed to read_apsim_soils}

\item{cleanup}{argument used to delete the file after download}
}
\value{
it creates a file with extension \sQuote{soils}
}
\description{
Retrieves soil data from the World Modeler global database and (optionally) saves it to a soils file
}
\examples{
\dontrun{
## Get soil profile properties for a single point
if(FALSE){
  tdir <- tempdir()
  sp1 <- get_worldmodeler_soil_profile(lonlat = c(-93, 42), wrt.dir = tdir, 
                                      filename = "temp_soils.soils")
  plot(sp1[[1]], property = "Carbon")
}
}
}
\author{
Brian Collins (University of Southern Queensland) and Fernando Miguez
}
