% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsimx_json.R
\name{grep_json_list}
\alias{grep_json_list}
\title{grep but for json list}
\usage{
grep_json_list(pattern, x, ignore.case = FALSE, search.depth = 10)
}
\arguments{
\item{pattern}{as in grep}

\item{x}{object (a list)}

\item{ignore.case}{as in grep}

\item{search.depth}{search depth for the list (to prevent endless search)}
}
\value{
It returns a list with the found object, the json path and the positions in the list.
}
\description{
recursive grep adapted for a json list
}
