% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsim_replace_soil_profile.R
\name{initialwater_parms}
\alias{initialwater_parms}
\title{Helper function to supply additional Initial Soil Water parameters}
\usage{
initialwater_parms(
  Depth = NA,
  Thickness = NA,
  InitialValues = NA,
  InitialPAWmm = NA,
  PercentFull = NA,
  RelativeTo = NA,
  FilledFromTop = NA,
  DepthWetSoil = NA
)
}
\arguments{
\item{Depth}{depth for soil layers (see APSIM documentation)}

\item{Thickness}{soil thickness for layers (either enter Depth or Thickness, but not both)}

\item{InitialValues}{initial values of soil water}

\item{InitialPAWmm}{Initial Plant Available Water in mm}

\item{PercentFull}{Percent full (0 - 100)}

\item{RelativeTo}{usually LL15}

\item{FilledFromTop}{either true or false}

\item{DepthWetSoil}{depth of wet soil in mm}
}
\description{
Creates a list with specific components for the Initial Soil Water module
}
