% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsim_xml.R
\name{inspect_apsim}
\alias{inspect_apsim}
\title{Inspect an .apsim (XML) file}
\usage{
inspect_apsim(file = "", src.dir = ".", node = c("Clock", "Weather",
  "Soil", "SurfaceOrganicMatter", "MicroClimate", "Crop", "Manager",
  "Other"), soil.child = c("Water", "OrganicMatter", "Analysis",
  "InitialWater", "Sample"), som.child = c("Pools", "Other"),
  parm = NULL, digits = 3)
}
\arguments{
\item{file}{file ending in .apsim (Classic) to be inspected (XML)}

\item{src.dir}{directory containing the .apsim file to be inspected; defaults to the current working directory}

\item{node}{either 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop', 'Manager' or 'Other'}

\item{soil.child}{specific soil component to be inspected}

\item{som.child}{specific surface organic matter component to be inspected ('Pools' or 'Other')}

\item{parm}{parameter to inspect when node = 'Other'}

\item{digits}{number of decimals to print (default 3)}
}
\value{
table with inspected parameters and values
}
\description{
inspect an XML apsim file. It does not replace the GUI, but it can save time by quickly checking parameters and values.
}
\details{
This is simply a script that prints the relevant parameters which are likely to need editing. It does not print all information from an .apsimx file.
}
\note{
node 'Manager' can be complicated and it is not guranteed to work.
}
\examples{
\dontrun{
extd.dir <- system.file("extdata", package = "apsimx")
## Testing using 'Millet'
inspect_apsim("Millet", src.dir = extd.dir, node = "Clock")
inspect_apsim("Millet", src.dir = extd.dir, node = "Weather") 
inspect_apsim("Millet", src.dir = extd.dir, node = "Soil", soil.child = "Water")
inspect_apsim("Millet", src.dir = extd.dir, node = "Soil", soil.child = "OrganicMatter")
inspect_apsim("Millet", src.dir = extd.dir, node = "Soil", soil.child = "Analysis")
inspect_apsim("Millet", src.dir = extd.dir, node = "Soil", soil.child = "InitialWater")
inspect_apsim("Millet", src.dir = extd.dir, node = "Soil", soil.child = "Sample")
inspect_apsim("Millet", src.dir = extd.dir, node = "SurfaceOrganicMatter")
inspect_apsim("Millet", src.dir = extd.dir, node = "Crop") 
inspect_apsim("Millet", src.dir = extd.dir, node = "Manager") 

## Testing with maize-soybean-rotation.apsim
inspect_apsim("maize-soybean-rotation", src.dir = extd.dir, node = "Clock")
}

}
