% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsimx.R
\name{inspect_apsimx}
\alias{inspect_apsimx}
\title{Inspect and .apsimx file}
\usage{
inspect_apsimx(file = "", src.dir = ".", node = c("Weather", "Soil",
  "SurfaceOrganicMatter", "MicroClimate", "Crop", "Manager"),
  soil.child = c("Water", "Nitrogen", "OrganicMatter", "Analysis",
  "InitialWater", "Sample"), som.child = c("Pools", "Other"),
  digits = 3)
}
\arguments{
\item{file}{file ending in .apsimx to be inspected}

\item{src.dir}{directory containing the .apsimx file to be inspected; defaults to the current working directory}

\item{node}{either 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop' or 'Manager'}

\item{soil.child}{specific soil component to be inspected}

\item{som.child}{specific soil organic matter component to be inspected}

\item{digits}{number of decimals to print}
}
\value{
table with inspected parameters and values
}
\description{
Similar structure to the 'edit_apsimx' file but aimed at a simple way
of inspecting a file
}
\note{
node 'Manager' can be complicated and it is not guranteed to work
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsimx_examples()
inspect_apsimx("Maize", src.dir = ex.dir, node = "Weather")        
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Water")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Nitrogen")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "OrganicMatter")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Analysis")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "InitialWater")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Sample")
inspect_apsimx("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Pools")
inspect_apsimx("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Other")
inspect_apsimx("Maize", src.dir = ex.dir, node = "MicroClimate")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Crop")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Manager")
}

}
