% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsimx_json.R
\name{inspect_apsimx_json}
\alias{inspect_apsimx_json}
\title{Inspect an .apsimx (JSON) file}
\usage{
inspect_apsimx_json(file = "", src.dir = ".", node = c("Clock",
  "Weather", "Soil", "SurfaceOrganicMatter", "MicroClimate", "Crop",
  "Manager", "Other"), soil.child = c("Water", "OrganicMatter",
  "Chemical", "Physical", "Analysis", "InitialWater", "Sample"),
  som.child = c("Pools", "Other"), parm = NULL, digits = 3)
}
\arguments{
\item{file}{file ending in .apsimx to be inspected (JSON)}

\item{src.dir}{directory containing the .apsimx file to be inspected; defaults to the current working directory}

\item{node}{either 'Clock', 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop', 'Manager' or 'Other'}

\item{soil.child}{specific soil component to be inspected. The options are: 'Water', 'OrganicMatter', 'Analysis', 'InitalWater', 'Sample'}

\item{som.child}{specific surface organic matter component to be inspected (not used)}

\item{parm}{parameter to refine the inspection of the 'manager' list('parm','position'), use 'NA' for all the positions. 'parm' can be a regular expression for partial matching.}

\item{digits}{number of decimals to print (default 3)}
}
\value{
table with inspected parameters and values
}
\description{
inspect a JSON apsimx file. It does not replace the GUI, but it can save time by quickly checking parameters and values.
}
\details{
This is simply a script that prints the relevant parameters which are likely to need editing. It does not print all information from an .apsimx file.
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsimx_examples()
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Clock") 
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Weather")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Soil", soil.child = "Water") 
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Soil", soil.child = "OrganicMatter")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Soil", soil.child = "Analysis")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Soil", soil.child = "InitialWater")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Soil", soil.child = "Sample")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "SurfaceOrganicMatter")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "MicroClimate")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Crop")
inspect_apsimx_json("Barley", src.dir = ex.dir, node = "Manager")

}

}
