% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsimx_xml.R
\name{inspect_apsimx_xml}
\alias{inspect_apsimx_xml}
\title{Inspect an .apsimx (XML) file}
\usage{
inspect_apsimx_xml(file = "", src.dir = ".", node = c("Clock",
  "Weather", "Soil", "SurfaceOrganicMatter", "MicroClimate", "Crop",
  "Manager"), soil.child = c("Water", "Nitrogen", "OrganicMatter",
  "Analysis", "InitialWater", "Sample"), som.child = c("Pools", "Other"),
  digits = 3)
}
\arguments{
\item{file}{file ending in .apsimx to be inspected (XML)}

\item{src.dir}{directory containing the .apsimx file to be inspected; defaults to the current working directory}

\item{node}{either 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop', 'Manager' or 'Other'}

\item{soil.child}{specific soil component to be inspected}

\item{som.child}{specific surface organic matter component to be inspected ('Pools' or 'Other')}

\item{digits}{number of decimals to print (default 3)}
}
\value{
table with inspected parameters and values
}
\description{
inspect an XML apsimx file. It does not replace the GUI, but it can save time by quickly checking parameters and values.
}
\details{
This is simply a script that prints the relevant parameters which are likely to need editing. It does not print all information from an .apsimx file.
}
\note{
node 'Manager' can be complicated and it is not guranteed to work.
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsimx_examples()
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Clock")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Weather") 
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Water")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Nitrogen")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "OrganicMatter")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Analysis")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "InitialWater")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Sample")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Pools")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Other")
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "MicroClimate") 
inspect_apsimx_xml("Maize", src.dir = ex.dir, node = "Crop") 
}

}
