% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_apsim.R
\name{mcmc.apsim.env}
\alias{mcmc.apsim.env}
\title{Environment to store data for apsim MCMC}
\usage{
mcmc.apsim.env
}
\value{
This is an environment, so nothing to return.
}
\description{
Environment which stores data for MCMC
}
\details{
Create an apsim environment for MCMC
}
