% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_apsimx.R
\name{mcmc.apsimx.env}
\alias{mcmc.apsimx.env}
\title{Environment to store data for apsimx MCMC}
\usage{
mcmc.apsimx.env
}
\value{
This is an environment, so nothing to return.
}
\description{
Environment which stores data for MCMC
}
\details{
Create an apsimx environment for MCMC
}
