% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{obsWheat}
\alias{obsWheat}
\title{Observed wheat phenology, LAI and biomass}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{Date}{ -date- date starting Oct 1 2016 and ending June 6 2017}
  \item{Wheat.Phenology.Stage}{ -numeric- phenology stage of wheat}
  \item{Wheat.Leaf.LAI}{ -numeric- Leaf Area Index}
  \item{Wheat.AboveGround.Wt}{-numeric- above ground biomass (g/m2)}
}
}
\source{
These are simulated data. For details see the APSIM documentation
}
\usage{
obsWheat
}
\description{
Artificial observed data for Wheat
}
\details{
A dataset containing the Date, phenology stage, LAI and above ground biomass for Wheat
}
\keyword{datasets}
