% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx_soil_profile.R
\name{plot.carbon.stocks}
\alias{plot.carbon.stocks}
\title{Plot carbon stocks (function in progress)}
\usage{
\method{plot}{carbon.stocks}(x, ...)
}
\arguments{
\item{x}{object of class \sQuote{soil_profile}}

\item{...}{optional argument, not used at the moment}
}
\value{
returns a graph
}
\description{
Plotting of carbon stocks based on an object of class \sQuote{soil_profile}
}
