% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{print.met}
\alias{print.met}
\title{Printer-friendly version of a metfile}
\usage{
\method{print}{met}(x, ...)
}
\arguments{
\item{x}{an R object of class \sQuote{met}}

\item{...}{additional printing arguments}
}
\description{
Print a met file in a friendly way
}
