% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{read_apsim_met}
\alias{read_apsim_met}
\title{Read in and check met file}
\usage{
read_apsim_met(file, src.dir = NULL, verbose = FALSE)
}
\arguments{
\item{file}{path to met file}

\item{src.dir}{optional source directory}

\item{quiet}{whether to suppress all messages and warnings}
}
\value{
an object of class 'met' with attributes
}
\description{
Read into R a met file, check validity and return an object of class 'met'
}
\details{
Read a met file into R

I will use S3 classes and store the additional information as attributes \cr
I use a more strict format than APSIM and reasing and writing will lose \cr
arbitrary constants
}
\examples{
\dontrun{
extd.dir <- system.file("extdata", package = "apsimx")
ames.met <- read_apsim_met("Ames.met", src.dir = extd.dir)
ames.met
}

}
