% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsim_replace_soil_profile.R
\name{soilorganicmatter_parms}
\alias{soilorganicmatter_parms}
\title{Helper function to supply additional Soil Organic Matter parameters}
\usage{
soilorganicmatter_parms(
  RootCN = NA,
  RootWt = NA,
  EnrACoeff = NA,
  EnrBCoeff = NA,
  OCUnits = NA
)
}
\arguments{
\item{RootCN}{Root Carbon:Nitrogen ratio (see APSIM documentation)}

\item{RootWt}{Root weight (see APSIM documentation)}

\item{EnrACoeff}{(see APSIM documentation)}

\item{EnrBCoeff}{(see APSIM documentation)}

\item{OCUnits}{Organic Carbon Units}
}
\description{
Creates a list with specific components for the Soil Organic Matter module
}
