% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsim_replace_soil_profile.R
\name{soilwat_parms}
\alias{soilwat_parms}
\title{Helper function to supply SoilWat parameters}
\usage{
soilwat_parms(
  SummerCona = NA,
  SummerU = NA,
  SummerDate = NA,
  WinterCona = NA,
  WinterU = NA,
  WinterDate = NA,
  DiffusConst = NA,
  DiffusSlope = NA,
  Salb = NA,
  CN2Bare = NA,
  CNRed = NA,
  CNCov = NA
)
}
\arguments{
\item{SummerCona}{see APSIM documentation}

\item{SummerU}{see APSIM documentation}

\item{SummerDate}{see APSIM documentation}

\item{WinterCona}{see APSIM documentation}

\item{WinterU}{see APSIM documentation}

\item{WinterDate}{see APSIM documentation}

\item{DiffusConst}{see APSIM documentation}

\item{DiffusSlope}{see APSIM documentation}

\item{Salb}{soil albedo (see APSIM documentation)}

\item{CN2Bare}{see APSIM documentation}

\item{CNRed}{see APSIM documentation}

\item{CNCov}{see APSIM documentation}
}
\value{
a \sQuote{list} with class \sQuote{soilwat_parms}
}
\description{
Creates a list with specific components for the SoilWat model
}
\details{
current documentation for APSIM 7.10 \url{https://www.apsim.info/documentation/model-documentation/soil-modules-documentation/soilwat/}
}
