% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{[.met}
\alias{[.met}
\title{Extract or Replace Part of a \sQuote{met} object}
\usage{
\method{[}{met}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of class \sQuote{met}}

\item{i}{index for rows}

\item{j}{index for columns}

\item{...}{additional arguments}

\item{drop}{result is coerced to lowest possible dimension if TRUE.}
}
\description{
Extract or Replace Part of a \sQuote{met} object
}
