% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_apsim_replace_soil_profile.R
\name{swim_parms}
\alias{swim_parms}
\title{Helper function to supply SWIM parameters}
\usage{
swim_parms(
  Salb = NA,
  CN2Bare = NA,
  CNRed = NA,
  CNCov = NA,
  KDul = NA,
  PSIDul = NA,
  VC = NA,
  DTmin = NA,
  DTmax = NA,
  MaxWaterIncrement = NA,
  SpaceWeightingFactor = NA,
  SoluteSpaceWeightingFactor = NA,
  Diagnostics = NA,
  SwimWaterTable_WaterTableDepth = NA,
  SwimSubsurfaceDrain_DrainDepth = NA,
  SwimSubsurfaceDrain_DrainSpacing = NA,
  SwimSubsurfaceDrain_DrainRadius = NA,
  SwimSubsurfaceDrain_Klat = NA,
  SwimSubsurfaceDrain_ImpermDepth = NA
)
}
\arguments{
\item{Salb}{see APSIM documentation}

\item{CN2Bare}{see APSIM documentation}

\item{CNRed}{see APSIM documentation}

\item{CNCov}{see APSIM documentation}

\item{KDul}{see APSIM documentation}

\item{PSIDul}{see APSIM documentation}

\item{VC}{see APSIM documentation}

\item{DTmin}{see APSIM documentation}

\item{DTmax}{see APSIM documentation}

\item{MaxWaterIncrement}{see APSIM documentation}

\item{SpaceWeightingFactor}{see APSIM documentation}

\item{SoluteSpaceWeightingFactor}{see APSIM documentation}

\item{Diagnostics}{see APSIM documentation}

\item{SwimWaterTable_WaterTableDepth}{see APSIM documentation}

\item{SwimSubsurfaceDrain_DrainDepth}{see APSIM documentation}

\item{SwimSubsurfaceDrain_DrainSpacing}{see APSIM documentation}

\item{SwimSubsurfaceDrain_DrainRadius}{see APSIM documentation}

\item{SwimSubsurfaceDrain_Klat}{see APSIM documentation}

\item{SwimSubsurfaceDrain_ImpermDepth}{see APSIM documentation}
}
\value{
a \sQuote{list} with class \sQuote{swim_parms}
}
\description{
Creates a list with specific components for the SWIM model
}
\details{
current documentation for APSIM 7.10 \url{https://www.apsim.info/documentation/model-documentation/soil-modules-documentation/swim3/}
}
