% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{tav_apsim_met}
\alias{tav_apsim_met}
\title{Calculates attribute amp for an object of class \sQuote{met}}
\usage{
tav_apsim_met(met, by.year = TRUE, na.rm = TRUE)
}
\arguments{
\item{met}{object of class \sQuote{met}}

\item{by.year}{whether to compute tav for each year and then average (default is TRUE)}

\item{na.rm}{whether to remove missing values (NAs). Default is TRUE.}
}
\value{
an object of class \sQuote{met} with a recalculation of annual mean temperature amplitude
}
\description{
This function can re-calculate annual mean temperature
for an object of class \sQuote{met}
}
