% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conv.R
\name{unit_conv}
\alias{unit_conv}
\title{performs common unit conversions}
\usage{
unit_conv(x, from, to)
}
\arguments{
\item{x}{input variable}

\item{from}{original units}

\item{to}{target units}
}
\value{
value of the input variable with new units
}
\description{
Function which performs common unit conversions
}
\details{
At the moment possible conversions are:
\itemize{

\item \sQuote{g/m2} to \sQuote{kg/ha}
\item \sQuote{kg/ha} to \sQuote{g/m2}

}
}
\examples{
\dontrun{
grain.yield.gm2 <- 600
grain.yield.kgha <- unit_conv(grain.yield.gm2, from = "g/m2", to = "kg/ha")
grain.yield.kgha
}

}
