% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsim_xml.R
\name{view_apsim_xml}
\alias{view_apsim_xml}
\title{View an APSIM Classic auxiliary (XML) file}
\usage{
view_apsim_xml(file, src.dir, viewer = c("json", "react"), ...)
}
\arguments{
\item{file}{file ending in .xml to be viewed.}

\item{src.dir}{directory containing the .xml file to be viewed; defaults to the current working directory}

\item{viewer}{either \dQuote{json} or \dQuote{react}.}

\item{...}{additional arguments passed to either \sQuote{jsonedit} or \sQuote{reactjson}.}
}
\value{
nothing
}
\description{
view an auxilliary XML apsim file.
}
\details{
view APSIM XML file
}
\examples{
 
\dontrun{
extd.dir <- system.file("extdata", package = "apsimx")
view_apsim_xml("Maize75.xml", src.dir = extd.dir)
}

}
