% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{wop.h}
\alias{wop.h}
\title{Wheat example optimization results plus Hessian}
\format{
An object of class 'optim_apsim' 
\describe{
  \item{ wop.h }{ wheat optimization results plus Hessian}
}
}
\source{
Result of running the examples in Parameter Optimization vignette with the added Hessian
}
\usage{
wop.h
}
\description{
Results from Wheat optimization example plus the Hessian
}
\keyword{datasets}
