% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{wop}
\alias{wop}
\title{Wheat example optimization results}
\format{
An object of class 'optim_apsim' 
\describe{
  \item{ wop }{ wheat optimization results }
}
}
\source{
Result of running the examples in Parameter Optimization vignette
}
\usage{
wop
}
\description{
Results from Wheat optimization example
}
\keyword{datasets}
