% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_met.R
\name{write_apsim_met}
\alias{write_apsim_met}
\title{Write a met file}
\usage{
write_apsim_met(met, wrt.dir = NULL, filename = NULL)
}
\arguments{
\item{met}{object of class met}

\item{path}{path where the file will be written}
}
\value{
a file ending in .met
}
\description{
Write an object of class 'met' to disk
}
\details{
Write a met file to disk. It takes an object of class 'met'

at the moment the read, write cycle will strip comments
}
\examples{
\dontrun{
extd.dir <- system.file("extdata", package = "apsimx")
ames.met <- read_apsim_met("Ames.met", src.dir = extd.dir)
ames.met
write_apsim_met(ames.met, wrt.dir = ".", filename = "Ames.met")
}

}
