% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_classic.R
\name{apsim}
\alias{apsim}
\title{Run an APSIM (7.x) \sQuote{Classic} simulation}
\usage{
apsim(
  file = "",
  src.dir = ".",
  silent = FALSE,
  value = c("all", "report", "none"),
  cleanup = FALSE
)
}
\arguments{
\item{file}{file name to be run (the extension .apsim is optional)}

\item{src.dir}{directory containing the .apsim file to be run (defaults to the current directory)}

\item{silent}{whether to print messages from apsim simulation}

\item{value}{how much output to return: \cr
option \sQuote{all} returns all components of the simulation; \cr
option \sQuote{report} returns only the \sQuote{main} report component; \cr
option \sQuote{none} does not create a data.frame but it generates the databases from APSIM-X}

\item{cleanup}{logical. Whether to delete the .out file generated by APSIM. Default is FALSE.}
}
\description{
Run apsim from R. It's for Windows only. It uses \sQuote{shell}.
}
\details{
Run an APSIM (7.x) \sQuote{Classic} Simulation

A valid apsim file can be run from within R. The main goal is to make running APSIM-X
simple, especially for large scale simulations or parameter optimization
}
\examples{
\dontrun{
## See function 'apsim_example' 
}

}
