% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_classic.R
\name{apsim}
\alias{apsim}
\title{Run an APSIM (7.x) \sQuote{Classic} simulation}
\usage{
apsim(
  file = "",
  src.dir = ".",
  silent = FALSE,
  value = c("report", "all", "none"),
  cleanup = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{file}{file name to be run (the extension .apsim is optional)}

\item{src.dir}{directory containing the .apsim file to be run (defaults to the current directory)}

\item{silent}{whether to print messages from apsim simulation}

\item{value}{how much output to return: \cr
option \sQuote{report} returns only the \sQuote{main} report component;
option \sQuote{all} returns all components of the simulation; \cr
option \sQuote{none} runs simulation but does not return a data frame.}

\item{cleanup}{logical. Whether to delete the .out and .sum files generated by APSIM. Default is FALSE.}

\item{simplify}{whether to return a single data frame when multiple simulations are present. If FALSE it will return a list.}
}
\value{
This function returns a data frame with APSIM output, but it depends on the argument \sQuote{value} above.
}
\description{
Run apsim from R. It's for Windows only. It uses \sQuote{shell}.
}
\details{
Run an APSIM (7.x) \sQuote{Classic} Simulation

A valid apsim file can be run from within R. The main goal is to make running APSIM-X
simple, especially for large scale simulations or parameter optimization
}
\examples{
\donttest{
## See function 'apsim_example' 
}

}
