% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\docType{data}
\name{apsimx.options}
\alias{apsimx.options}
\title{Environment which stores APSIM-X options}
\format{
An object of class \code{environment} of length 9.
}
\usage{
apsimx.options
}
\value{
This is an environment, not a function, so nothing is returned.
}
\description{
Environment which can store the path to the executable, warning settings and 
             where examples are located.
             Creating an environment avoids the use of global variables or other similar practices
             which would have possible undesriable consequences.
}
\details{
Environment which stores APSIM-X options
}
\examples{
\donttest{
names(apsimx.options)
apsimx_options(exe.path = "some-new-path-to-executable")
apsimx.options$exe.path
}

}
\keyword{datasets}
