% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{apsimx}
\alias{apsimx}
\title{Run an APSIM-X simulation}
\usage{
apsimx(file = "", src.dir = ".", silent = FALSE, value = c("all",
  "report", "none"), cleanup = c(0, 1, 2))
}
\arguments{
\item{file}{file name to be run (the extension .apsimx is optional)}

\item{src.dir}{directory containing the .apsimx file to be run (defaults to the current directory)}

\item{silent}{whether to print messages from apsim simulation}

\item{value}{how much output to return: \cr
option 'all' returns all components of the simulation; \cr
option 'report' returns only the 'main' report component; \cr
option 'none' does not create a data.frame but it generates the databases from APSIM-X}

\item{cleanup}{level of file cleanup: \cr
0 does nothing; \cr
1 deletes the .db created by APSIM-X; \cr
2 is not implemented yet.}
}
\description{
Run apsimx from R. It uses 'system' and it attempts to be platform independent.
}
\details{
Run an APSIM-X Simulation

A valid apsimx file can be run from within R. The main goal is to make running APSIM-X
simple, especially for large scale simulations or parameter optimization
}
\examples{
\dontrun{
## See function 'apsimx_example' 
}

}
