% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{apsimx_options}
\alias{apsimx_options}
\title{Setting some options for the package}
\usage{
apsimx_options(exe.path = NA, examples.path = NA)
}
\arguments{
\item{exe.path}{path to the executable}

\item{examples.path}{path to the examples}
}
\description{
Used to set the path to the APSIM-X executable.
}
\details{
Set APSIM-X options
}
\note{
It is possible that APSIM-X is installed in some alternative location other than the 
      defaults ones. Guessing this can be difficult and then the auto_detect functions might
      fail.
}
\examples{
\dontrun{
names(apsimx.options)
apsimx_options(exe.path = "some-new-path-to-executable")
apsimx.options$exe.path
}
}
