% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{apsimx_options}
\alias{apsimx_options}
\title{Setting some options for the package}
\usage{
apsimx_options(exe.path = NA, examples.path = NA,
  warn.versions = TRUE)
}
\arguments{
\item{exe.path}{path to apsim executable}

\item{examples.path}{path to apsim examples}

\item{warn.versions}{logical. warning if multiple versions of APSIM-X are detected.}
}
\description{
Set the path to the APSIM-X executable, examples and warning suppression.
}
\details{
Set apsimx options
}
\note{
It is possible that APSIM-X is installed in some alternative location other than the 
      defaults ones. Guessing this can be difficult and then the auto_detect functions might
      fail. Also, if multiple versions of APSIM-X are installed apsimx will choose the newest
      one but it will issue a warning. Suppress the warning by setting warn.versions = FLASE.
}
\examples{
\dontrun{
names(apsimx.options)
apsimx_options(exe.path = "some-new-path-to-executable")
apsimx.options$exe.path
}
}
