% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_apsimx.R
\name{inspect_apsimx}
\alias{inspect_apsimx}
\title{Inspect an .apsimx file (XML or JSON)}
\usage{
inspect_apsimx(file = "", src.dir = ".", node = c("Clock", "Weather",
  "Soil", "SurfaceOrganicMatter", "MicroClimate", "Crop", "Manager",
  "Other"), soil.child = c("Water", "OrganicMatter", "Analysis",
  "Chemical", "InitialWater", "Sample"), som.child = c("Pools", "Other"),
  parm = NULL, digits = 3)
}
\arguments{
\item{file}{file ending in .apsimx to be inspected}

\item{src.dir}{directory containing the .apsimx file to be inspected; defaults to the current working directory}

\item{node}{either 'Clock', 'Weather', 'Soil', 'SurfaceOrganicMatter', 'MicroClimate', 'Crop', 'Manager' or 'Other'}

\item{soil.child}{specific soil component to be inspected. For example, 'Water' or 'OrganicMatter'.}

\item{som.child}{specific soil organic matter component to be inspected ('Pools' or 'Other')}

\item{parm}{parameter to inspect when node = 'Other' (XML only)}

\item{digits}{number of decimals to print}
}
\value{
table with inspected parameters and values
}
\description{
inspect either an XML or JSON apsimx file. It does not replace the GUI, but it can save time by quickly checking parameters and values.
}
\details{
Similar structure to the 'edit_apsimx' file but aimed at a simple way
of inspecting a file

The inspection has many elements that are hard coded, it is not a literal inspection tool for all aspects of the file. For this use a text editor.
}
\note{
node 'Manager' can be complicated and it is not guranteed to work. First version of a file which supports
both XML and JSON files. This is simply a script that prints the relevant parameters which are likely to need editing. It does not print all information from an .apsimx file.
}
\examples{
\dontrun{
ex.dir <- auto_detect_apsimx_examples()
inspect_apsimx("Maize", src.dir = ex.dir, node = "Clock")        
inspect_apsimx("Maize", src.dir = ex.dir, node = "Weather")        
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Water")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "OrganicMatter")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Chemical")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "InitialWater")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Soil", soil.child = "Sample")
inspect_apsimx("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Pools")
inspect_apsimx("Maize", src.dir = ex.dir, node = "SurfaceOrganicMatter", som.child = "Other")
inspect_apsimx("Maize", src.dir = ex.dir, node = "MicroClimate")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Crop")
inspect_apsimx("Maize", src.dir = ex.dir, node = "Manager")
}

}
