% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_apsim.R
\docType{data}
\name{mcmc.apsim.env}
\alias{mcmc.apsim.env}
\title{Environment to store data for apsim MCMC}
\format{
An object of class \code{environment} of length 0.
}
\usage{
mcmc.apsim.env
}
\description{
Environment which stores data for MCMC
}
\details{
Create an apsim environment for MCMC
}
\keyword{datasets}
