% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_apsimx.R
\docType{data}
\name{mcmc.apsimx.env}
\alias{mcmc.apsimx.env}
\title{Environment to store data for apsimx MCMC}
\format{
An object of class \code{environment} of length 0.
}
\usage{
mcmc.apsimx.env
}
\description{
Environment which stores data for MCMC
}
\details{
Create an apsimx environment for MCMC
}
\keyword{datasets}
