% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsim_classic.R
\name{read_apsim_all}
\alias{read_apsim_all}
\title{Read all APSIM generated .out files in a directory}
\usage{
read_apsim_all(
  filenames,
  src.dir = ".",
  value = "report",
  date.format = "\%d/\%m/\%Y",
  simplify = TRUE,
  silent = FALSE
)
}
\arguments{
\item{filenames}{names of files to be read}

\item{src.dir}{source directory where files are located}

\item{value}{either \sQuote{report}, \sQuote{user-defined} or \sQuote{all} (not implemented at the moment)}

\item{date.format}{format for adding \sQuote{Date} column}

\item{simplify}{whether to return a single data frame or a list.}

\item{silent}{whether to issue warnings or suppress them}
}
\value{
returns a data frame or a list depending on the argument \sQuote{simplify} above.
}
\description{
Like \code{\link{read_apsim}}, but it can read many .out files in a directory. 
It will read all of them unless these are filtered using a regular expression as an argument
to \sQuote{value}.
}
\details{
Read all APSIM generated .out files in a directory
}
\note{
Warning: very simple function at the moment, not optimized for memory or speed.
}
