% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{read_apsimx}
\alias{read_apsimx}
\title{Read APSIM-X generated .db files}
\usage{
read_apsimx(file = "", src.dir = ".", value = c("report", "all"))
}
\arguments{
\item{file}{file name}

\item{src.dir}{source directory where file is located}

\item{value}{either \sQuote{report} (data.frame) or \sQuote{all} (list)}
}
\description{
read SQLite databases created by APSIM-X runs. One file at a time.
}
\details{
Read APSIM-X generated .db files
}
