% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{read_apsimx_all}
\alias{read_apsimx_all}
\title{Read all APSIM-X generated .db files in a directory}
\usage{
read_apsimx_all(src.dir = ".", value = c("report", "all"))
}
\arguments{
\item{src.dir}{source directory where files are located}

\item{value}{either \sQuote{report} or \sQuote{all} (only \sQuote{report} implemented at the moment)}
}
\description{
Like \code{\link{read_apsimx}}, but it reads all .db files in a directory.
}
\details{
Read all APSIM-X generated .db files in a directory
}
\note{
Warning: very simple function at the moment, not optimized for memory or speed.
}
