% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssurgo2sp.R
\name{ssurgo2sp}
\alias{ssurgo2sp}
\title{Take in \acronym{SSURGO} csv files and create a soil profile}
\usage{
ssurgo2sp(mapunit = NULL, component = NULL, chorizon = NULL,
  nmapunit = 1, nsoil = 1, xout = NULL, soil.bottom = 200,
  method = c("constant", "linear"), nlayers = 10)
}
\arguments{
\item{mapunit}{mapunit \acronym{SSURGO} file}

\item{component}{component \acronym{SSURGO} file}

\item{chorizon}{chorizon \acronym{SSURGO} file}

\item{nmapunit}{number of mapunits to select}

\item{nsoil}{number of soil components (within a mapunit) to consider}

\item{xout}{vector for interpolation and extrapolation}

\item{soil.bottom}{bottom of the soil profile}

\item{method}{method used for interpolation (see \code{\link{approx}})}

\item{nlayers}{number of soil layers to generate}
}
\value{
a list with soil profile matrices with length equal to nsoil
}
\description{
Utility function to convert \acronym{SSURGO} data to soil profile
}
\details{
Download the data from \acronym{SSURGO} using the \sQuote{FedData} package \cr
This will generate csv files \sQuote{chorizon}, \sQuote{component} and \sQuote{mapunit}, \cr
but also many other files which are not needed for creating a soil profile.
}
\examples{
\dontrun{
extd.dir <- system.file("extdata", package = "apsimx")

chorizon <- read_csv(paste0(extd.dir,"/ISUAG/SSURGO/ISUAG_SSURGO_chorizon.csv"))
component <- read_csv(paste0(extd.dir,"/ISUAG/SSURGO/ISUAG_SSURGO_component.csv"))
mapunit <- read_csv(paste0(extd.dir,"/ISUAG/SSURGO/ISUAG_SSURGO_mapunit.csv"))

## Using default 'constant' method
sp.c <- ssurgo2sp(mapunit = mapunit, 
                 component = component, 
                 chorizon = chorizon)
require(ggplot2)
ggplot(data = sp.c, aes(y = -Depth, x = Carbon)) + 
geom_point() + 
  geom_path() + 
  ylab("Soil Depth (cm)") + xlab("Organic Matter (percent)") +
  ggtitle("method = constant")
  
## Using 'linear' method
sp.l <- ssurgo2sp(mapunit = mapunit, 
                 component = component, 
                 chorizon = chorizon, method = "linear")
                 
ggplot(data = sp.l, aes(y = -Depth, x = Carbon)) + 
geom_point() + 
  geom_path() + 
  ylab("Soil Depth (cm)") + xlab("Organic Matter (percent)") +
  ggtitle("Method linear")
  
}
}
