% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx.R
\name{xargs_apsimx}
\alias{xargs_apsimx}
\title{Provide extra arguments for APSIM-X}
\usage{
xargs_apsimx(
  verbose = FALSE,
  csv = FALSE,
  merge.db.files = FALSE,
  list.simulations = FALSE,
  list.referenced.filenames = FALSE,
  single.threaded = FALSE,
  cpu.count = -1L,
  simulation.names = FALSE
)
}
\arguments{
\item{verbose}{Write detailed messages to stdout when a simulation starts/finishes.}

\item{csv}{Export all reports to .csv files.}

\item{merge.db.files}{Merge multiple .db files into a single .db file.}

\item{list.simulations}{List simulation names without running them.}

\item{list.referenced.filenames}{List all files that are referenced by an .apsimx file(s).}

\item{single.threaded}{Run all simulations sequentially on a single thread.}

\item{cpu.count}{(Default: -1) Maximum number of threads/processes to spawn for running simulations.}

\item{simulation.names}{Only run simulations if their names match this regular expression.}
}
\description{
This provides additinoal command line arguments when running the model
}
\details{
Extra arguments for running APSIM-X
}
